+++ rouver/html.py
@@ -4,14 +4,16 @@ from rouver.types import BadArgumentsDict
 
 
 def http_status_page(status: HTTPStatus, *,
-                     message: str = "", content: str = "") -> str:
+                     html_message: str = "", html_content: str = "") -> str:
     """Create an HTML error page for a given status code.
 
-    WARNING: The arguments "message" and "content" are not safe! Their content
-    will be pasted into the page as is. Do not use with unsanitized data!
+    WARNING: The arguments "html_message" and "html_content" are not safe!
+    Their content will be pasted into the page as is. Do not use with
+    unsanitized data!
     """
-    paragraph = "\n        <p>{}</p>".format(message) if message else ""
-    content = content + "\n" if content else ""
+    paragraph = \
+        "\n        <p>{}</p>".format(html_message) if html_message else ""
+    content = html_content + "\n" if html_content else ""
     return """<!DOCTYPE html>
 <html>
     <head>
@@ -26,23 +28,25 @@ def http_status_page(status: HTTPStatus, *,
 
 def created_at_page(url: str) -> str:
     message = 'Created at <a href="{0}">{0}</a>.'.format(url)
-    return http_status_page(HTTPStatus.CREATED, message=message)
+    return http_status_page(HTTPStatus.CREATED, html_message=message)
 
 
 def temporary_redirect_page(url: str) -> str:
     message = 'Please see <a href="{0}">{0}</a>.'.format(url)
-    return http_status_page(HTTPStatus.TEMPORARY_REDIRECT, message=message)
+    return http_status_page(
+        HTTPStatus.TEMPORARY_REDIRECT, html_message=message)
 
 
 def see_other_page(url: str) -> str:
     message = 'Please see <a href="{0}">{0}</a>.'.format(url)
-    return http_status_page(HTTPStatus.SEE_OTHER, message=message)
+    return http_status_page(HTTPStatus.SEE_OTHER, html_message=message)
 
 
 def bad_arguments_page(arguments: BadArgumentsDict) -> str:
     content = bad_arguments_list(arguments)
     return http_status_page(
-        HTTPStatus.BAD_REQUEST, message="Invalid arguments:", content=content)
+        HTTPStatus.BAD_REQUEST, html_message="Invalid arguments:",
+        html_content=content)
 
 
 def bad_arguments_list(arguments: BadArgumentsDict) -> str: